改bug

+++ src/main/java/com/example/modules/wall/controller/PostCommentController.java
@@ -3,22 +3,18 @@ package com.example.modules.wall.controller;
 
 import com.example.modules.user.pojo.po.User;
 import com.example.modules.user.service.UserService;
-import com.example.modules.user.utils.Consts;
 import com.example.modules.wall.entity.dto.PostCommentFromViewDTO;
 import com.example.modules.wall.entity.po.Post;
 import com.example.modules.wall.entity.po.PostComment;
-
-import java.util.ArrayList;
-import java.util.List;
-
 import com.example.modules.wall.entity.vo.PostCommentVO;
 import com.example.modules.wall.service.PostCommentService;
 import com.example.modules.wall.service.PostService;
 import com.example.utils.R;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/wall/comment")
@@ -76,13 +72,17 @@ public class PostCommentController {
             // redis获取当前用户id
             String userId = userServiceImpl.getTokenUser().getId();
             PostComment comment = postCommentServiceImpl.getComment(commentId);
-            postCommentServiceImpl.deleteComment(commentId);
             String postId = comment.getPostId();
             Post post = postServiceImpl.getPostById(postId);
-            post.setCommentCount(post.getCommentCount() - 1);
-            // 更新评论数量
-            postServiceImpl.updatePostById(post);
-            return R.success(null);
+            if(userId.equals(comment.getUserId())){
+                postCommentServiceImpl.deleteComment(commentId);
+                post.setCommentCount(post.getCommentCount() - 1);
+                // 更新评论数量
+                postServiceImpl.updatePostById(post);
+                return R.success(null, "成功", 200);
+            }else{
+                return R.success(null, "没有权限", 200);
+            }
         }catch (Exception e){
             return R.error();
         }
